requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.49
botocore==1.27.49
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
decorator==5.1.1
Django==3.2
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.13.0
django-mongodb-engine @ git+https://github.com/django-nonrel/mongodb-engine@1314786975d3d17fb1a2e870dfba004158a5e431
django-mongodb-engine-py3==0.6.0.1
django-rest-framework-mongoengine==3.4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.6
dnspython==2.2.1
dynaconf==3.1.9
geographiclib==2.0
geopy==2.3.0
haversine==2.7.0
hvac==0.11.2
idna==3.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
mongoengine==0.24.2
numpy==1.22.1
pandas==1.4.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.9.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.2.4
threadpoolctl==3.1.0
tzdata==2022.1
urllib3==1.26.9
validators==0.20.0
win32-setctime==1.1.0
